fix: write to the text files in the correct order.
The while loop cannot be asynchronous, so the i++ was done to correctly loop through data. However, fs.appendFile is asynchronous, so all of the write actions are queued immediately until the loop resolves. Node cannot guarantee that these async actions will be done in a particular order, so you must wait until one write is complete before queuing the next write.

The simplest way to solve this is to use appendFileSync, which is blocking -- the loop will not increment until the current write is complete.

Also starting clearing the files when the application runs.