feat(streams): read and write data to files using streams.
Streams are a powerful, non-blocking feature of node.js that tend to be useful for large data and handling http requests, files, and databases. Streams are incompatible with promises, and they are event listeners that require callbacks to respond to their events such as 'finish' for when stream writing is ended.

This keeps the functionality identical, but it uses streams and callbacks instead of promises in applicable areas.

Note that I kept the callback name 'resolve' for download data to showcase that there is nothing special about the resolve callback name for Promises. This functions identically to our own callback we pass in, the one that logs 'done!'