fix: clean up the codebase
These are non-functional changes that are largely based on my personal standards. I used the prettier tool to format all of the JavaScript files. I renamed some of the files and symbol names for more clarity and standards. I implemented error handling.

I also added a configuration file with the relevent filenames so these are not hard-coded in the app and can be changed as needed. Otherwise, it's possible to mistype one of these names which can lead to errors that are hard to track (see: Magic numbers/Magic strings).